Chained validators POC
